backend/requirements.txt
fastapi==0.122.0
uvicorn==0.38.0
sqlalchemy==2.0.44
pydantic==2.12.5
psycopg2-binary==2.9.11
boto3==1.41.5
python-dotenv==1.0.1
requests==2.32.3
PyMuPDF==1.24.9
opencv-python-headless==4.10.0.84
numpy==1.26.4
python-multipart==0.0.9
pdf2image==1.16.3
Pillow==10.4.0
